Make Makefile a bit smarter at finding dependencies include path
This uses a combination of testing for known locations and pkg-config
to try to locate the right -I magic for the AWS SDK, pkcs11.h and
json-c. Additionally, they can be specified explicitly as arguments
to 'make'.

This makes "make" work out of the box on my ubuntu 20.04 and AL2.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>