Try to deduce the module install path
This uses pkg-config looking for p11-kit-1 or nss. I have no provision
for opencryptoki since it doesn't come with a pkg-config file in
the distros I have, so in absence of both of the above, one needs to
pass an explicit path, which is better than hard wiring one that
only works on x86_64 with debian based distros.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>