Fix build on some modern Ubuntu systems
Modern Ubuntu systems place shared libraries in a subdirectory of
`/usr/lib` indicated by the output of `gcc -dumpmachine`; on modern
x86_64 systems this is `/usr/lib/x86_64-linux-gnu`. The Makefile did not
account for this case, therefore requiring modifications before building
on Ubuntu systems.

This commit also modifies the Makefile to respect an `AWS_SDK_LIB_PATH`
that was overridden on the command line.

Signed-Off-By: Dan Fuhry <dan@fuhry.com>