Allow setting AWS_SDK_LIB_PATH instead of detecting it
This allows a little more control over the build in case the SDK isn't
installed in a typial way. In particular, this helps with the case where
the SDK is vendored with aws-kms-pkcs11 and built at the same time. In
that case, running "make clean" will fail unless the SDK is already
built.